Move declaration type to Statement
